+++ plugins/ci/pkg/models/models.go
@@ -1,11 +1,14 @@
 package models
 
 import (
+	"encoding/json"
 	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
+
+	"github.com/sirupsen/logrus"
 )
 
 // ScoreOutOfBoundsMessage is the message for the error when the score returned by Insights is out of bounds.
@@ -104,17 +107,59 @@ const (
 	AzureDevops   CIRunnerVal = "azure-devops"
 )
 
+type RegistryCredential struct {
+	Domain   string `yaml:"domain"`
+	Username string `yaml:"username"`
+	Password string `yaml:"password"`
+}
+
+func (rc RegistryCredential) String() string {
+	return fmt.Sprintf("domain: %s, username: %s, password: %s", rc.Domain, rc.Username, strings.Map(func(r rune) rune { return '*' }, rc.Password))
+}
+
+type RegistryCredentials []RegistryCredential
+
+func (rc RegistryCredentials) Validate() error {
+	// make sure there is no duplicated domain
+	domains := map[string]struct{}{}
+	for _, v := range rc {
+		domains[v.Domain] = struct{}{}
+	}
+	if len(rc) > len(domains) {
+		return errors.New("duplicated domains found in registry credentials list")
+	}
+	return nil
+}
+
+func (rc RegistryCredentials) FindCredentialForImage(imageName string) *RegistryCredential {
+	parts := strings.Split(imageName, "/")
+	domain := "docker.io"
+	if len(parts) == 3 {
+		// quay.io/fedora/httpd:version1.0
+		domain = parts[0]
+	}
+
+	for _, v := range rc {
+		if v.Domain == domain {
+			return &v
+		}
+	}
+
+	return nil
+}
+
 type optionConfig struct {
-	SetExitCode            bool        `yaml:"setExitCode"`
-	BaseBranch             string      `yaml:"baseBranch"`
-	NewActionItemThreshold int         `yaml:"newActionItemThreshold"`
-	SeverityThreshold      string      `yaml:"severityThreshold"`
-	TempFolder             string      `yaml:"tempFolder"`
-	Hostname               string      `yaml:"hostname"`
-	Organization           string      `yaml:"organization"`
-	JUnitOutput            string      `yaml:"junitOutput"`
-	RepositoryName         string      `yaml:"repositoryName"`
-	CIRunner               CIRunnerVal `yaml:"-"`
+	SetExitCode            bool                `yaml:"setExitCode"`
+	BaseBranch             string              `yaml:"baseBranch"`
+	NewActionItemThreshold int                 `yaml:"newActionItemThreshold"`
+	SeverityThreshold      string              `yaml:"severityThreshold"`
+	TempFolder             string              `yaml:"tempFolder"`
+	Hostname               string              `yaml:"hostname"`
+	Organization           string              `yaml:"organization"`
+	JUnitOutput            string              `yaml:"junitOutput"`
+	RepositoryName         string              `yaml:"repositoryName"`
+	RegistryCredentials    RegistryCredentials `yaml:"-"`
+	CIRunner               CIRunnerVal         `yaml:"-"`
 }
 
 type imageConfig struct {
@@ -208,7 +253,7 @@ func (c *Configuration) SetPathDefaults() {
 //
 // it should respect the order:
 // - config. file content > env. variables > default
-func (c *Configuration) SetDefaults() {
+func (c *Configuration) SetDefaults() error {
 	c.Options.CIRunner = CIRunnerVal(strings.TrimSpace(os.Getenv("CI_RUNNER"))) // only set via env. variable
 
 	if c.Options.BaseBranch == "" {
@@ -257,8 +302,23 @@ func (c *Configuration) SetDefaults() {
 		c.Reports.Trivy.SkipManifests = &falsehood
 	}
 	if c.Reports.TFSec.Enabled == nil {
-		c.Reports.TFSec.Enabled = &falsehood
+		c.Reports.TFSec.Enabled = &truth
 	}
+
+	registryCredentialsJSON := strings.TrimSpace(os.Getenv("REGISTRY_CREDENTIALS")) // only set via env. variable
+	if registryCredentialsJSON != "" {
+		var registryCredentials RegistryCredentials
+		err := json.Unmarshal([]byte(registryCredentialsJSON), &registryCredentials)
+		if err != nil {
+			return fmt.Errorf("could not parse registry credentials: %w", err)
+		}
+		if err := registryCredentials.Validate(); err != nil {
+			return fmt.Errorf("registryCredentials is not valid: %w", err)
+		}
+		c.Options.RegistryCredentials = registryCredentials
+		logrus.Infof("loaded %d registry credentials", len(registryCredentials))
+	}
+	return nil
 }
 
 // CheckForErrors checks to make sure the configuration is valid